Refactor reporters into hpp/cpp
