+++ CMakeLists.txt
@@ -56,6 +56,7 @@ set(SNITCH_INCLUDES
     ${PROJECT_SOURCE_DIR}/include/snitch/snitch.hpp
     ${PROJECT_SOURCE_DIR}/include/snitch/snitch_append.hpp
     ${PROJECT_SOURCE_DIR}/include/snitch/snitch_capture.hpp
+    ${PROJECT_SOURCE_DIR}/include/snitch/snitch_catch2_xml.hpp
     ${PROJECT_SOURCE_DIR}/include/snitch/snitch_cli.hpp
     ${PROJECT_SOURCE_DIR}/include/snitch/snitch_concepts.hpp
     ${PROJECT_SOURCE_DIR}/include/snitch/snitch_console.hpp
@@ -69,6 +70,7 @@ set(SNITCH_INCLUDES
     ${PROJECT_SOURCE_DIR}/include/snitch/snitch_macros_constexpr.hpp
     ${PROJECT_SOURCE_DIR}/include/snitch/snitch_macros_exceptions.hpp
     ${PROJECT_SOURCE_DIR}/include/snitch/snitch_macros_misc.hpp
+    ${PROJECT_SOURCE_DIR}/include/snitch/snitch_macros_reporter.hpp
     ${PROJECT_SOURCE_DIR}/include/snitch/snitch_macros_test_case.hpp
     ${PROJECT_SOURCE_DIR}/include/snitch/snitch_macros_utility.hpp
     ${PROJECT_SOURCE_DIR}/include/snitch/snitch_macros_warnings.hpp
@@ -86,15 +88,16 @@ set(SNITCH_INCLUDES
 set(SNITCH_SOURCES_INDIVIDUAL
     ${PROJECT_SOURCE_DIR}/src/snitch_append.cpp
     ${PROJECT_SOURCE_DIR}/src/snitch_capture.cpp
+    ${PROJECT_SOURCE_DIR}/src/snitch_catch2_xml.cpp
     ${PROJECT_SOURCE_DIR}/src/snitch_cli.cpp
     ${PROJECT_SOURCE_DIR}/src/snitch_console.cpp
     ${PROJECT_SOURCE_DIR}/src/snitch_error_handling.cpp
     ${PROJECT_SOURCE_DIR}/src/snitch_main.cpp
     ${PROJECT_SOURCE_DIR}/src/snitch_matcher.cpp
     ${PROJECT_SOURCE_DIR}/src/snitch_registry.cpp
-    ${PROJECT_SOURCE_DIR}/src/snitch_reporters.cpp
     ${PROJECT_SOURCE_DIR}/src/snitch_section.cpp
     ${PROJECT_SOURCE_DIR}/src/snitch_string_utility.cpp
+    ${PROJECT_SOURCE_DIR}/src/snitch_teamcity.cpp
     ${PROJECT_SOURCE_DIR}/src/snitch_test_data.cpp)
 
 if (SNITCH_UNITY_BUILD)
