+++ meson.build
@@ -8,6 +8,7 @@ include_dirs = include_directories('.', 'include')
 headers = files('include/snitch/snitch.hpp',
                 'include/snitch/snitch_append.hpp',
                 'include/snitch/snitch_capture.hpp',
+                'include/snitch/snitch_catch2_xml.hpp',
                 'include/snitch/snitch_cli.hpp',
                 'include/snitch/snitch_concepts.hpp',
                 'include/snitch/snitch_console.hpp',
@@ -21,6 +22,7 @@ headers = files('include/snitch/snitch.hpp',
                 'include/snitch/snitch_macros_constexpr.hpp',
                 'include/snitch/snitch_macros_exceptions.hpp',
                 'include/snitch/snitch_macros_misc.hpp',
+                'include/snitch/snitch_macros_reporter.hpp',
                 'include/snitch/snitch_macros_test_case.hpp',
                 'include/snitch/snitch_macros_utility.hpp',
                 'include/snitch/snitch_macros_warnings.hpp',
@@ -36,15 +38,16 @@ headers = files('include/snitch/snitch.hpp',
 
 sources = files('src/snitch_append.cpp',
                'src/snitch_capture.cpp',
+               'src/snitch_catch2_xml.cpp',
                'src/snitch_cli.cpp',
                'src/snitch_console.cpp',
                'src/snitch_error_handling.cpp',
                'src/snitch_main.cpp',
                'src/snitch_matcher.cpp',
                'src/snitch_registry.cpp',
-               'src/snitch_reporters.cpp',
                'src/snitch_section.cpp',
                'src/snitch_string_utility.cpp',
+               'src/snitch_teamcity.cpp',
                'src/snitch_test_data.cpp')
 
 if get_option('unity_build')
